requirements.txt
pandas
chromadb
sentence-transformers
google-generativeai
python-dotenv